// sanity.types.ts
export declare const internalGroqTypeReferenceTo: unique symbol;

// Source: schema.json
export type FranchisePackage = {
  _id: string;
  _type: "franchisePackage";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  name?: string;
  price?: string;
  priceNote?: string;
  features?: Array<string>;
  featured?: boolean;
  order?: number;
};

export type SanityImageAssetReference = {
  _ref: string;
  _type: "reference";
  _weak?: boolean;
  [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
};
// ...
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export type ALL_BRANCHES_QUERY_RESULT = Array<{
  _id: string;
  zone: string | null;
  name: string | null;
  city: string | null;
  province: string | null;
  address: string | null;
  tables: number | null;
  tags: Array<string> | null;
  openHour: string | null;
  closeHour: string | null;
  mapsUrl: string | null;
  photo: {
    asset?: SanityImageAssetReference;
    media?: unknown;
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
  } | null;
}>;